+++ src/Stryker.Core/Stryker.Core.UnitTest/TestRunners/VsTestRunnerPoolTests.cs
@@ -10,6 +10,7 @@ using Stryker.Core.Exceptions;
 using Stryker.Core.Initialisation;
 using Stryker.Core.Mutants;
 using Stryker.Core.Options;
+using Stryker.Core.TestRunners;
 using Stryker.Core.TestRunners.VsTest;
 using Xunit;
 
@@ -147,7 +148,7 @@ namespace Stryker.Core.UnitTest.TestRunners
             var action = () =>  runner.TestMultipleMutants(SourceProjectInfo, null, new[] { Mutant }, null);
             action.ShouldThrow<GeneralStrykerException>();
             // the test will always end in a crash, VsTestRunner should retry at least a few times
-            mockVsTest.Verify(m => m.RunTestsWithCustomTestHostAsync(It.IsAny<IEnumerable<string>>(),
+            mockVsTest.Verify(m => m.RunTestsWithCustomTestHost(It.IsAny<IEnumerable<string>>(),
                 It.IsAny<string>(), It.IsAny<TestPlatformOptions>(),
                 It.IsAny<ITestRunEventsHandler>(),
                 It.IsAny<IStrykerTestHostLauncher>()), Times.AtLeast(3));
@@ -189,11 +190,11 @@ namespace Stryker.Core.UnitTest.TestRunners
             var mockVsTest = BuildVsTestRunnerPool(new StrykerOptions(), out var runner);
             var defaultTimeOut = VsTestRunner.VsTestExtraTimeOutInMs;
             VsTestRunner.VsTestExtraTimeOutInMs = 100;
-            // the test session will hung twice
+            // the test session will freeze twice
             SetupFrozenTestRun(mockVsTest, 2);
             runner.TestMultipleMutants(SourceProjectInfo, new TimeoutValueCalculator(0, 10,9), new[] { Mutant }, null);
             VsTestRunner.VsTestExtraTimeOutInMs = defaultTimeOut;
-            mockVsTest.Verify(m => m.RunTestsWithCustomTestHostAsync(It.IsAny<IEnumerable<string>>(),
+            mockVsTest.Verify(m => m.RunTestsWithCustomTestHost(It.IsAny<IEnumerable<string>>(),
                 It.IsAny<string>(), It.IsAny<TestPlatformOptions>(),
                 It.IsAny<ITestRunEventsHandler>(),
                 It.IsAny<IStrykerTestHostLauncher>()), Times.Exactly(3));
@@ -205,7 +206,7 @@ namespace Stryker.Core.UnitTest.TestRunners
             var mockVsTest = BuildVsTestRunnerPool(new StrykerOptions(), out var runner);
             var defaultTimeOut = VsTestRunner.VsTestExtraTimeOutInMs;
             // the test session will end properly, but VsTest will hang
-            // it will be recyled
+            // it will be recycled
             SetupFrozenVsTest(mockVsTest, 3);
             VsTestRunner.VsTestExtraTimeOutInMs = 100;
             runner.TestMultipleMutants(SourceProjectInfo, new TimeoutValueCalculator(0, 10,9), new[] { Mutant }, null);
@@ -575,7 +576,7 @@ namespace Stryker.Core.UnitTest.TestRunners
         }
 
         // this verifies that tests missing any coverage information are
-        // flagged as to be tested used against every mutants
+        // flagged as to be tested used against every mutant
         [Fact]
         public void MarkSuspiciousTests()
         {
@@ -609,6 +610,35 @@ namespace Stryker.Core.UnitTest.TestRunners
 
             var testResult = BuildCoverageTestResult("T0", new[] { "0;", "" });
             var other = BuildCoverageTestResult("T1", new[] { "", "" });
+            SetupMockCoverageRun(mockVsTest, new[] { testResult, other });
+
+
+            var analyzer = new CoverageAnalyser(options);
+            analyzer.DetermineTestCoverage(SourceProjectInfo, runner, new[] { Mutant, OtherMutant }, TestGuidsList.NoTest());
+
+            OtherMutant.CoveringTests.Count.ShouldBe(0);
+            Mutant.CoveringTests.Count.ShouldBe(1);
+        }
+
+        // this verifies extra test results (without any coverage info) are properly handled
+        // are properly handled
+        [Fact]
+        public void HandleExtraTestResult()
+        {
+            var options = new StrykerOptions
+            {
+                OptimizationMode = OptimizationModes.CoverageBasedTest
+            };
+
+            var mockVsTest = BuildVsTestRunnerPool(options, out var runner);
+
+            var testResult = BuildCoverageTestResult("T0", new[] { "0;", "" });
+            var other =  new TestResult(FindOrBuildCase("T0"))
+            {
+                DisplayName = "T0",
+                Outcome = TestOutcome.Passed,
+                ComputerName = "."
+            };
             SetupMockCoverageRun(mockVsTest, new[] { testResult, other });