+++ src/Stryker.Core/Stryker.Core/Reporters/Html/RealTime/Events/SseEvent.cs
@@ -1,4 +1,4 @@
-﻿using System.Text.Json;
+using System.Text.Json;
 
 namespace Stryker.Core.Reporters.Html.RealTime.Events;
 
@@ -7,9 +7,8 @@ public class SseEvent<T>
     public SseEventType Event { get; init; }
     public T Data { get; init; }
 
-    public string Serialize() =>
-        $@"
-            event:{Event.Serialize()}
-            data:{JsonSerializer.Serialize(Data, new JsonSerializerOptions { PropertyNamingPolicy = JsonNamingPolicy.CamelCase })}
-        ".Replace(" ", "");
+    public string Serialize() => $@"
+event:{Event.Serialize()}
+data:{JsonSerializer.Serialize(Data, new JsonSerializerOptions { PropertyNamingPolicy = JsonNamingPolicy.CamelCase })}
+";
 }